# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Inventario",
        "version": "1.0"
    },
    "servers": [
        {
            "url": true
        }
    ],
    "paths": {
        "api/productosimple": {
            "post": {
                "summary": "Guardar producto simple",
                "operationId": "App\\Http\\Controllers\\Simple\\ProductosimplePostController::__invoke",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "path",
                        "description": "nombre del producto simple",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "costo",
                        "in": "path",
                        "description": "costo del producto simple",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "precio",
                        "in": "path",
                        "description": "precio de venta del producto simple",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "cantidad",
                        "in": "path",
                        "description": "cantidad del producto simple",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "preparacion",
                        "in": "path",
                        "description": "si el producto es de preparación (SI-NO)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Se guardo correctamente."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "api/productosimple/salida/{nombre}/{cantidad}": {
            "get": {
                "summary": "Get project information",
                "description": "Returns project data",
                "operationId": "App\\Http\\Controllers\\Simple\\ProductosimpleSalidaController::__invoke",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "path",
                        "description": "nombre del producto simple",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cantidad",
                        "in": "path",
                        "description": "cantidad del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}